Essentials/pyproject.toml
[project]
name = "essentials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "requests",
]
